classApiwithGraphql/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==3.2.4
django-filter==24.3
django-graphql-auth==0.3.16
django-graphql-jwt==0.4.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
promise==2.3
psycopg2==2.9.10
PyJWT==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
Rx==1.6.3
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing-extensions==4.13.1
